requirements.txt
certifi
future
mkl-fft
mkl-random
mkl-service
numpy
Pillow
pyglet
six
wincertstore